requirements.txt
playwright
pandas
xlsxwriter
tk
pyyaml
pytest
pytest-mock
hypothesis
colorama
openpyxl
pillow